requirements.txt
httpx>=0.27
pydantic>=2.5
pydantic-settings>=2.1
pyyaml>=6.0
python-pptx>=0.6.23
pysrt>=1.1.2
webvtt-py>=0.5
langdetect>=1.0.9
rich>=13.7
click>=8.1
fastapi>=0.109
uvicorn[standard]>=0.27
python-multipart>=0.0.6
python-docx>=1.1
zhconv>=1.4.3
# Phase A: Markdown, JSON (stdlib), YAML
mistune>=3.0
ruamel.yaml>=0.18
# Phase B: PO/POT, XLIFF (lxml already used by docx), XML (lxml)
polib>=1.2.0
lxml>=5.0
# Phase C: HTML
beautifulsoup4>=4.12
